Move dictionary indexing tests to its own file in indexing module
